% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectar_metages.R
\name{conectar_metages}
\alias{conectar_metages}
\title{Conectar con la base de datos METAGES}
\usage{
conectar_metages(driver = NULL, entorno = c("prod", "test"))
}
\arguments{
\item{driver}{Nombre del driver ODBC a utilizar. Por defecto se usa
\code{"MySQL ODBC 9.4 Unicode Driver"}, pero puede variar segun el sistema
operativo y la instalacion local.
\strong{Solo se admiten drivers ODBC Unicode}. Los drivers ANSI no estan soportados.
Para listar los drivers disponibles desde R: \code{odbc::odbcListDrivers()}.
En sistemas donde el driver por defecto no funcione, el usuario
deberá especificar uno alternativo mediante el argumento \code{driver}.}

\item{entorno}{Entorno de MetaGES al que conectarse. Puede ser \code{"prod"}
(valor por defecto) o \code{"test"}. Cada entorno utiliza sus propias variables
de ambiente para el host, el túnel SSH y la contraseña de la base de datos.}
}
\value{
Una lista con dos elementos:
\describe{
\item{con}{Conexión DBI a la base de datos.}
\item{tunnel}{Objeto del túnel SSH.}
}
}
\description{
Establece una conexión con la base de datos METAGES mediante un
túnel SSH y una conexión ODBC.
}
\details{
La función depende de variables de entorno previamente configuradas
(por ejemplo, host, claves SSH y credenciales de base de datos).
Para mas informacion sobre la configuracion de las variables de entorno, ver
\href{https://gbifes.github.io/metages-toolkit/articles/guia-uso-dev.html#configuracion-de--renviron-para-acceder-a-metages}{Documentación técnica de metagesToolkit}

La conexión se realiza en dos pasos:
\enumerate{
\item Apertura de un túnel SSH.
\item Conexión a la base de datos vía \pkg{DBI} y \pkg{ODBC}.
}
La base de datos MetaGES utiliza codificacion UTF-8 y contiene caracteres
internacionales (acentos, caracteres cientificos, etc.).

Por este motivo, \code{conectar_metages()} \strong{solo permite el uso de drivers ODBC Unicode}.
Los drivers ODBC ANSI no soportan correctamente UTF-8 y pueden
provocar errores silenciosos o corrupcion de texto.

Si no se especifica un driver, la función utiliza el driver Unicode por
defecto. Si dicho driver no está instalado en el sistema, la conexión
fallará.

Para instalar un driver ODBC Unicode para MySQL, consulte la página oficial:
\url{https://dev.mysql.com/downloads/connector/odbc/}
}
