% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertar_tabla_colecciones.R
\name{crear_flextable_colecciones}
\alias{crear_flextable_colecciones}
\title{Crear flextable de colecciones}
\usage{
crear_flextable_colecciones(tabla)
}
\arguments{
\item{tabla}{Un \code{data.frame} devuelto por
\code{crear_tabla_colecciones()}.}
}
\value{
Un objeto \code{flextable} listo para ser insertado en un documento Word.
}
\description{
Convierte una tabla de colecciones en un objeto \code{flextable}, añadiendo
hipervínculos en las columnas de institución y colección, ajustando las
cabeceras y eliminando columnas auxiliares no visibles.
}
