% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_mapa_entidades.R
\name{crear_mapa_entidades}
\alias{crear_mapa_entidades}
\title{Crear mapa de entidades en el Registro facetado según publicación en GBIF}
\usage{
crear_mapa_entidades(tipo_coleccion = NULL)
}
\arguments{
\item{tipo_coleccion}{Tipo de entidad a representar. Uno de
\code{"coleccion"}, \code{"base de datos"} o \code{NULL}.
Si es \code{NULL}, se muestran todas.}
}
\value{
Un objeto \code{ggplot}.
}
\description{
Genera un mapa de España que muestra la localización de las entidades
incluidas en el Registro de Colecciones de GBIF.ES. El mapa se facetiza según el
estado de publicación en GBIF (\code{publica_en_gbif}), separando visualmente
las entidades que publican de aquellas que no.
}
\details{
Las entidades se representan mediante puntos de tamaño fijo. Para evitar
ocultamientos completos entre entidades cercanas, se aplica un jitter
espacial muy leve. Los nombres que aparecen en el mapa corresponden a las
regiones, y se sitúan en el centroide empírico de las entidades de cada
región y panel, manteniendo coherencia cartográfica.
}
