% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_colecciones_mapa.R
\name{extraer_colecciones_mapa}
\alias{extraer_colecciones_mapa}
\title{Extraer colecciones desde METAGES para mapas}
\usage{
extraer_colecciones_mapa(
  shift = c(5, 6),
  cerrar_conexion = FALSE,
  cerrar_tunel = FALSE
)
}
\arguments{
\item{shift}{Vector numérico de longitud 2. Se usa en el procesamiento para
desplazar los datos de canarias y coincidir con \code{get_basemap_es()}.}

\item{cerrar_conexion}{Si \code{TRUE}, cierra la conexión DB al finalizar.}

\item{cerrar_tunel}{Si \code{TRUE}, cierra el túnel/proceso al finalizar.}
}
\value{
Invisiblemente, un data.frame/tibble con los datos de colecciones
listos para usar en \code{crear_mapa()}.
}
\description{
Abre conexión a METAGES (vía \code{conectar_metages()}), ejecuta una consulta
(\verb{SELECT * FROM colecciones c}) y devuelve un data.frame depurado para mapas.
}
