% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basemap_es.R
\name{get_basemap_es}
\alias{get_basemap_es}
\title{Obtener basemap de España (con Canarias desplazadas)}
\usage{
get_basemap_es(shift = c(5, 6))
}
\arguments{
\item{shift}{Vector numérico de longitud 2 con el desplazamiento aplicado
a Canarias (x, y).}
}
\value{
Invisiblemente, una lista con:
\describe{
\item{vecinos}{Objeto sf con países vecinos.}
\item{ES_fixed}{Objeto sf con España (Península+Canarias desplazadas).}
\item{bb_fixed}{Bounding box (\code{sf::st_bbox}) de \code{ES_fixed}.}
\item{bb_can}{Bounding box (\code{sf::st_bbox}) de Canarias desplazadas.}
\item{shift}{El vector \code{shift} usado para desplazar canarias.}
}
}
\description{
Descarga geometrías de países vecinos y España con \code{giscoR}, recorta
y desplaza Canarias para crear un basemap “compacto” para mapas.
}
