% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectar_metages.R
\name{conectar_metages}
\alias{conectar_metages}
\title{Conectar con la base de datos METAGES}
\usage{
conectar_metages(driver = "MySQL ODBC 9.4 Unicode Driver")
}
\arguments{
\item{driver}{Nombre del driver ODBC a utilizar. Por defecto se usa
\code{"MySQL ODBC 9.4 Unicode Driver"}, pero puede variar según el sistema
operativo y la instalación local. Para listar los drivers disponibles:
\code{odbc::odbcListDrivers()}.

En sistemas donde el driver por defecto no funcione, el usuario
deberá especificar uno alternativo mediante el argumento \code{driver}.}
}
\value{
Una lista con dos elementos:
\describe{
\item{con}{Conexión DBI a la base de datos.}
\item{tunnel}{Objeto del túnel SSH.}
}
}
\description{
Establece una conexión con la base de datos METAGES mediante un
túnel SSH y una conexión ODBC.
}
\details{
La función depende de variables de entorno previamente configuradas
(por ejemplo, host, claves SSH y credenciales de base de datos).
Para mas informacion sobre la configuracion de las variables de entorno, ver
\href{https://gbifes.github.io/metages-toolkit/articles/guia-uso-dev.html#configuracion-de--renviron-para-acceder-a-metages}{Documentación técnica de metagesToolkit}

La conexión se realiza en dos pasos:
\enumerate{
\item Apertura de un túnel SSH.
\item Conexión a la base de datos vía \pkg{DBI} y \pkg{ODBC}.
}
}
