% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_barplot_top_colecciones_pub.R
\name{crear_barplot_top_colecciones_pub}
\alias{crear_barplot_top_colecciones_pub}
\title{Crear gráfico de top colecciones por número de registros}
\usage{
crear_barplot_top_colecciones_pub(rds_path)
}
\arguments{
\item{rds_path}{Ruta al archivo \code{.rds} que contiene el mapa de colecciones.}
}
\value{
Un objeto \code{ggplot}.
}
\description{
Genera un gráfico de barras horizontales con las 10 colecciones
con mayor número de registros que publican en GBIF, a partir
de un objeto leído desde un archivo \code{.rds} interno de metagesToolkit.
}
