% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_mapa.R
\name{crear_mapa}
\alias{crear_mapa}
\title{Crear mapa de colecciones METAGES}
\usage{
crear_mapa(
  data = data,
  basemap,
  legend_params = NULL,
  tipo_coleccion = NULL,
  disciplina = NULL,
  subdisciplina = NULL,
  publican = NULL,
  facet = NULL
)
}
\arguments{
\item{data}{data.frame con las columnas necesarias para el mapa. Se
genera automaticamente con \code{extraer_colecciones_mapa()}}

\item{basemap}{Lista devuelta por \code{get_basemap_es()}.}

\item{legend_params}{Lista de parametros de leyenda. Si \code{NULL}, se calcula con
\code{compute_legend_params(data)}.}

\item{tipo_coleccion}{\code{coleccion}, \verb{base de datos} o \code{NULL}.}

\item{disciplina}{\code{Zoológica}, \code{Botánica}, \code{Paleontológica}, \code{Mixta}, \code{Microbiológica}, \code{Micológica} o \code{NULL}.}

\item{subdisciplina}{\code{Vertebrados}, \code{Invertebrados}, \verb{Invertebrados y vertebrados}, \code{Plantas}, \code{Hongos}, \code{Algas} o \code{NULL}.}

\item{publican}{\code{TRUE}, \code{FALSE} o \code{NULL}.}

\item{facet}{Nombre de columna (string) para facetar o \code{NULL}.}
}
\value{
Invisiblemente, una lista con:
\describe{
\item{plot}{Objeto \code{ggplot}.}
\item{data_map}{data.frame con los datos tras filtros.}
}
}
\description{
Genera un mapa con puntos sobre un basemap, aplicando filtros opcionales y
un facet opcional. Devuelve tanto el \code{ggplot} como los datos filtrados.
}
\details{
La funcion asume que los datos de entrada contienen coordenadas
geograficas y métricas asociadas a colecciones.
}
