% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_mapa_entidades.R
\name{crear_mapa_entidades}
\alias{crear_mapa_entidades}
\title{Crear mapa de entidades en el Registro facetado segun publicacion en GBIF}
\usage{
crear_mapa_entidades(tipo_coleccion = NULL)
}
\arguments{
\item{tipo_coleccion}{Tipo de entidad a representar. Uno de
\code{"coleccion"}, \code{"base de datos"} o \code{NULL}.
Si es \code{NULL}, se muestran todas.}
}
\value{
Un objeto \code{ggplot}.
}
\description{
Genera un mapa del país que muestra la localizacion de las entidades
incluidas en el Registro de Colecciones de GBIF.ES. El mapa se facetiza segun el
estado de publicacion en GBIF (\code{publica_en_gbif}), separando visualmente
las entidades que publican de aquellas que no.
}
\details{
Las entidades se representan mediante puntos fijos. Para evitar
ocultamientos completos entre entidades cercanas, se aplica un jitter
espacial muy leve. Los nombres que aparecen en el mapa corresponden a las
regiones, y se situan en el centroide empirico de las entidades de cada
region y panel, manteniendo coherencia cartografica.
}
