% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_mapa_simple.R
\name{crear_mapa_simple}
\alias{crear_mapa_simple}
\title{Mapa METAGES (version simple para uso)}
\usage{
crear_mapa_simple(
  tipo_coleccion = NULL,
  disciplina = NULL,
  subdisciplina = NULL,
  publican = NULL,
  facet = NULL
)
}
\arguments{
\item{tipo_coleccion}{\code{colección}, \verb{base de datos} o \code{NULL}.}

\item{disciplina}{\code{Zoológica}, \code{Botánica}, \code{Paleontológica}, \code{Mixta}, \code{Microbiológica}, \code{Micológica} o \code{NULL}.}

\item{subdisciplina}{\code{Vertebrados}, \code{Invertebrados}, \verb{Invertebrados y vertebrados}, \code{Plantas}, \code{Hongos}, \code{Algas} o \code{NULL}.}

\item{publican}{\code{TRUE}, \code{FALSE} o \code{NULL}.}

\item{facet}{Nombre de columna (string) para facetar o \code{NULL}.}
}
\value{
Invisiblemente, una lista con:
\describe{
\item{plot}{Objeto \code{ggplot}.}
\item{data_map}{data.frame con los datos tras filtros.}
}
}
\description{
Wrapper de alto nivel sobre \code{crear_mapa()} que fija la infraestructura
y expone solo filtros y facet.
}
