% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_informe.R
\name{render_informe}
\alias{render_informe}
\title{Renderiza el informe METAGES}
\usage{
render_informe()
}
\value{
Crea varias carpetas y archivos de cache, figuras y un documento .docx.
}
\description{
Renderiza el archivo Quarto \code{informe.qmd} incluido en el paquete.
}
